Separate gann_abc from gann_waves
