+++ tbot/indicators/qte/__init__.py
@@ -3,6 +3,7 @@
 from .gann_abc import GannABC
 from .gann_bar import GannBar
 from .gann_dir import GannDir
+from .gann_waves import GannWaves
 from .gann_with_hoagie import GannWithHoagie
 
-__all__ = ["GannABC", "GannBar", "GannDir", "GannWithHoagie"]
+__all__ = ["GannABC", "GannBar", "GannDir", "GannWaves", "GannWithHoagie"]
